requirements.txt
# Core dependencies for speech-to-text application
faster-whisper>=1.0.0
sounddevice>=0.4.6
pynput>=1.7.6
PyQt5>=5.15.0
numpy>=1.24.0
pyyaml>=6.0
pyperclip>=1.8.2

# Additional utilities
pyaudio>=0.2.13

# Build tools (for creating exe)
pyinstaller>=6.0.0
